Minor updates to CAMUS dataset (#292)
Update camus dataset - resize_inputs functionality
